+++ core/src/test/scala/de/welt/contentapi/core/client/services/aws/s3/S3ClientTest.scala
@@ -1,10 +1,10 @@
 package de.welt.contentapi.core.client.services.aws.s3
 
+import java.io.ByteArrayInputStream
 import java.nio.charset.Charset
 
 import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.model.{GetObjectRequest, S3Object, S3ObjectInputStream}
-import org.apache.commons.io.IOUtils
 import org.mockito.{ArgumentMatchers, Mockito}
 import org.scalatestplus.play.PlaySpec
 
@@ -16,13 +16,15 @@ class S3ClientTest extends PlaySpec {
     val service: S3Client = new S3Client(s3)
   }
 
+
   "S3Client" should {
     "pass requests to the underlying amazon client" in new Scope {
       Mockito.when(s3.getObject(ArgumentMatchers.any[GetObjectRequest]))
         .thenReturn(response)
+
+      //noinspection ScalaStyle
       Mockito.when(response.getObjectContent)
-        .thenReturn(new S3ObjectInputStream(
-          IOUtils.toInputStream("hello, world!", Charset.defaultCharset()), null))
+        .thenReturn(new S3ObjectInputStream(new ByteArrayInputStream("hello, world!".getBytes(Charset.defaultCharset())), null))
       val s3_response = service.get("foo", "bar")
 
       s3_response mustBe Some("hello, world!")