+++ pressed/src/test/scala/de/welt/contentapi/pressed/client/converter/Raw2ApiSiteBuildingTest.scala
@@ -1,163 +1,183 @@
 package de.welt.contentapi.pressed.client.converter
 
-import de.welt.contentapi.core.models.{ApiAsset, ApiElement, ApiReference}
-import de.welt.contentapi.pressed.models.{ApiSiteBuildingConfiguration}
-import de.welt.contentapi.raw.models
 import de.welt.contentapi.raw.models._
 import de.welt.testing.TestHelper.raw.channel.{emptyWithId, emptyWithIdAndChildren}
+import org.mockito.Mockito
 import org.scalatestplus.play.PlaySpec
 
+//noinspection ScalaStyle
 class Raw2ApiSiteBuildingTest extends PlaySpec {
 
-  val converter: RawToApiConverter = new RawToApiConverter(new InheritanceCalculator())
+  val rawToApiConverter: RawToApiConverter = new RawToApiConverter(new InheritanceCalculator())
 
   //noinspection ScalaStyle
   trait SiteBuildingTreeScope {
+    def spyRawToApiConverter = Mockito.spy(rawToApiConverter)
 
-    // @formatter:off
+    /** no sitebuilding */
+    val channel = emptyWithIdAndChildren(200, Seq())
+    channel.config = RawChannelConfiguration()
 
-    /**
-      * *S* = nonEmpty site building
-      * *s* = site building equals the default constructor
-      * *M* = is explicit master
-      * *m* = is implicit master, because 1st level child of root
-      *
-      *
-      *         (    0[root]   *S*    )
-      *         |            \        \
-      *      (10 *SM*)    (20 *M*)   (30 *m*)
-      *        |              \
-      *      (100)         (200 *S*)
-      *       |                \
-      *     (1000 *s*)     (2000 *S*)
-      */
-
-    // @formatter:on
-
-    /** 1XXX */
-    val node1000 = emptyWithId(1000)
-    node1000.config = RawChannelConfiguration(siteBuilding = Some(RawChannelSiteBuilding()))
-
-    val node1000WithDeletedSitebuildingConf = emptyWithId(1000)
-    node1000WithDeletedSitebuildingConf.config = RawChannelConfiguration(siteBuilding = Some(RawChannelSiteBuilding(fields = Some(Map.empty[String, String]))))
-
-    val node100 = emptyWithIdAndChildren(100, Seq(node1000, node1000WithDeletedSitebuildingConf))
-
-    val node10 = emptyWithIdAndChildren(10, Seq(node100))
-    node10.config = RawChannelConfiguration(siteBuilding = Some(RawChannelSiteBuilding(
-      fields = Some(Map("key1" -> "value2", "key2" -> "value2")),
+    /** master and has everything to inherit */
+    val master = emptyWithIdAndChildren(20, Seq(channel))
+    master.config = RawChannelConfiguration(siteBuilding = Some(RawChannelSiteBuilding(
+      fields = Some(Map(
+        "header_slogan" -> "Header Slogan",
+        "sponsoring_slogan" -> "Sponsoring Slogan"
+      )),
       sub_navigation = Some(Seq(RawSectionReference(Some("Label"), Some("/Path/")))),
-      elements = Some(Seq(
-        RawElement(
-          id = RawChannelElement.IdDefault,
-          `type` = "mood",
-          assets = Some(List(
-            RawAsset(
-              `type` = "image",
-              fields = Some(Map("key1" -> "value2", "key2" -> "value2"))
-            )
-          ))
-        )
-      ))
-    )), master = true)
-
-    /** 2XXX */
-    val node2000 = emptyWithId(200)
-    node2000.config = RawChannelConfiguration(siteBuilding = None)
-
-    val node200 = emptyWithIdAndChildren(200, Seq(node2000))
-    node200.config = RawChannelConfiguration(siteBuilding = Some(RawChannelSiteBuilding(fields = Some(Map("header_label" -> "Label")))))
-
-    val node20 = emptyWithIdAndChildren(20, Seq(node200))
-    node20.config = RawChannelConfiguration(master = true, siteBuilding = None)
-
-    /** 3X */
-    val node30 = emptyWithId(30)
-
-    /** root */
-    val root = emptyWithIdAndChildren(0, Seq(node10, node20, node30))
-    root.config = RawChannelConfiguration(siteBuilding = Some(RawChannelSiteBuilding(
-      fields = Some(Map("header_slogan" -> "Slogan")),
-      sub_navigation = Some(Seq(RawSectionReference(Some("Label"), Some("/Path/"))))
+      elements = Some(Seq(RawElement(id = "element id", `type` = "element type")))
     )))
+
+    /** fake root tree */
+    val root = emptyWithIdAndChildren(0, Seq(master))
     root.updateParentRelations()
-
     import de.welt.contentapi.core.models.testImplicits.pathUpdater
-
     root.updatePaths()
   }
 
-  "Sitebuilding Conversion WITHOUT master inheritance" must {
+  "calculateSiteBuilding()" should {
 
-    "be chosen if channel defines a sitebuilding field" in {
-      val channel = emptyWithId(1000)
-      channel.config = RawChannelConfiguration(siteBuilding = Some(models.RawChannelSiteBuilding(fields = Some(Map("header_label" -> "Label")))))
-      converter.calculateSiteBuilding(channel) mustBe Some(ApiSiteBuildingConfiguration(fields = Some(Map("header_label" -> "Label"))))
+    "look for a master and inherit `sponsoring_` values if channel.isMasterInheritanceEligible is true because no `sponsoring_` values are present" in new SiteBuildingTreeScope {
+      channel.config = RawChannelConfiguration(siteBuilding = Some(RawChannelSiteBuilding(
+        fields = Some(Map(
+          "header_slogan" -> "Header Slogan"
+        )),
+        sub_navigation = Some(Seq(RawSectionReference(Some("Label"), Some("/Path/")))),
+        elements = Some(Seq(RawElement(id = "element id", `type` = "element type")))
+      )))
+
+      spyRawToApiConverter.calculateSiteBuilding(channel).get.unwrappedFields.get("sponsoring_slogan") mustBe Some("Sponsoring Slogan")
+      Mockito.verify(spyRawToApiConverter, Mockito.atMostOnce()).calculateMasterChannel(channel)
     }
 
-    "be chosen if channel defines a sitebuilding element" in {
+    "look for a master and inherit `header_` values if channel.isMasterInheritanceEligible is true because no `header` values are present" in new SiteBuildingTreeScope {
+      channel.config = RawChannelConfiguration(siteBuilding = Some(RawChannelSiteBuilding(
+        fields = Some(Map(
+          "sponsoring_slogan" -> "Sponsoring Slogan"
+        )),
+        sub_navigation = Some(Seq(RawSectionReference(Some("Label"), Some("/Path/")))),
+        elements = Some(Seq(RawElement(id = "element id", `type` = "element type")))
+      )))
 
-      val channel = emptyWithId(1000)
-      channel.config = RawChannelConfiguration(siteBuilding = Some(models.RawChannelSiteBuilding(
-        elements = Some(Seq(
-          RawElement(
-            id = RawChannelElement.IdDefault,
-            `type` = "mood",
-            assets = Some(List(
-              RawAsset(
-                `type` = "image",
-                fields = Some(Map("key1" -> "value2", "key2" -> "value2"))
-              )
-            ))
-          )
-        )))
-      ))
-      converter.calculateSiteBuilding(channel) mustBe Some(ApiSiteBuildingConfiguration(
-        elements = Some(Seq(
-          ApiElement(
-            id = RawChannelElement.IdDefault,
-            `type` = "mood",
-            assets = Some(List(
-              ApiAsset(
-                `type` = "image",
-                fields = Some(Map("key1" -> "value2", "key2" -> "value2"))
-              )
-            ))
-          )
-        ))
-      ))
+      spyRawToApiConverter.calculateSiteBuilding(channel).get.unwrappedFields.get("header_slogan") mustBe Some("Header Slogan")
+      Mockito.verify(spyRawToApiConverter, Mockito.atMostOnce()).calculateMasterChannel(channel)
     }
 
-    "be chosen if channel defines a sub navigation (SectionReferences)" in {
-      val channel = emptyWithId(1000)
-      channel.config = RawChannelConfiguration(siteBuilding = Some(RawChannelSiteBuilding(sub_navigation = Some(Seq(RawSectionReference(path = Some("/path/")))))))
-      converter.calculateSiteBuilding(channel) mustBe Some(ApiSiteBuildingConfiguration(sub_navigation = Some(Seq(ApiReference(href = Some("/path/"))))))
+    "look for a master and inherit `elements` if channel.isMasterInheritanceEligible is true because no elements are present" in new SiteBuildingTreeScope {
+      channel.config = RawChannelConfiguration(siteBuilding = Some(RawChannelSiteBuilding(
+        fields = Some(Map(
+          "header_slogan" -> "Header Slogan",
+          "sponsoring_slogan" -> "Sponsoring Slogan"
+        )),
+        sub_navigation = Some(Seq(RawSectionReference(Some("Label"), Some("/Path/")))),
+        elements = None
+      )))
+
+      spyRawToApiConverter.calculateSiteBuilding(channel).get.unwrappedElements.size mustBe 1
+      Mockito.verify(spyRawToApiConverter, Mockito.atMostOnce()).calculateMasterChannel(channel)
     }
+
+    "look for a master and inherit `sub_navigation` if channel.isMasterInheritanceEligible is true because no sub_navigation is present" in new SiteBuildingTreeScope {
+      channel.config = RawChannelConfiguration(siteBuilding = Some(RawChannelSiteBuilding(
+        fields = Some(Map(
+          "header_slogan" -> "Header Slogan",
+          "sponsoring_slogan" -> "Sponsoring Slogan"
+        )),
+        sub_navigation = None,
+        elements = Some(Seq(RawElement(id = "element id", `type` = "element type")))
+      )))
+
+      spyRawToApiConverter.calculateSiteBuilding(channel).get.unwrappedSubNavigation.size mustBe 1
+      Mockito.verify(spyRawToApiConverter, Mockito.atMostOnce()).calculateMasterChannel(channel)
+    }
+
+    "never look for a master and inherit values if channel.isMasterInheritanceEligible is false" in new SiteBuildingTreeScope {
+      channel.config = RawChannelConfiguration(siteBuilding = Some(RawChannelSiteBuilding(
+        fields = Some(Map(
+          "header_slogan" -> "Header Slogan",
+          "sponsoring_slogan" -> "Sponsoring Slogan"
+        )),
+        sub_navigation = Some(Seq(RawSectionReference(Some("Label"), Some("/Path/")))),
+        elements = Some(Seq(RawElement(id = "element id", `type` = "element type")))
+      )))
+
+      spyRawToApiConverter.calculateSiteBuilding(channel).isDefined mustBe true
+      Mockito.verify(spyRawToApiConverter, Mockito.never()).calculateMasterChannel(channel)
+    }
+
+
+
   }
 
-  "Site Building Conversion WITH master inheritance" must {
+  "mergeSitebuildings()" should {
 
-    "be chosen if values are equal to constructor defaults (empty site building configuration)" in new SiteBuildingTreeScope {
-      converter.calculateSiteBuilding(node1000) mustBe converter.calculateSiteBuilding(node10)
+    "add `header_` fields from master if channel itself has no `header_` fields" in {
+      val channel = RawChannelSiteBuilding(fields = Some(Map(
+        "sponsoring_slogan" -> "Winamp - it really whips the llama's ass",
+        "sponsoring_hidden" -> "true"
+      )))
+
+      val master = RawChannelSiteBuilding(fields = Some(Map(
+        "header_logo" -> "zukunftsfond",
+        "header_hidden" -> "true"
+      )))
+      rawToApiConverter.mergeSitebuildings(channel, master) mustBe RawChannelSiteBuilding(fields = Some(Map(
+        "sponsoring_slogan" -> "Winamp - it really whips the llama's ass",
+        "sponsoring_hidden" -> "true",
+        "header_logo" -> "zukunftsfond",
+        "header_hidden" -> "true"
+      )))
     }
 
-    "be chosen if values are empty (deleted site building configuration)" in new SiteBuildingTreeScope {
-      converter.calculateSiteBuilding(node1000WithDeletedSitebuildingConf) mustBe converter.calculateSiteBuilding(node10)
+    "add `header_` fields from master if channel itself has `header_hidden = false` as only `header` field" in {
+      val channel = RawChannelSiteBuilding(fields = Some(Map(
+        "sponsoring_slogan" -> "Winamp - it really whips the llama's ass",
+        "sponsoring_hidden" -> "true",
+        "header_hidden" -> "false"
+      )))
+
+      val master = RawChannelSiteBuilding(fields = Some(Map(
+        "header_logo" -> "zukunftsfond",
+        "header_hidden" -> "true"
+      )))
+      rawToApiConverter.mergeSitebuildings(channel, master) mustBe RawChannelSiteBuilding(fields = Some(Map(
+        "sponsoring_slogan" -> "Winamp - it really whips the llama's ass",
+        "sponsoring_hidden" -> "true",
+        "header_logo" -> "zukunftsfond",
+        "header_hidden" -> "true"
+      )))
     }
+
+    "add `sponsoring_` fields from master if channel itself has no `sponsoring_` fields" in {
+      val channel = RawChannelSiteBuilding(fields = Some(Map(
+        "header_logo" -> "zukunftsfond"
+      )))
+
+      val master = RawChannelSiteBuilding(fields = Some(Map(
+        "header_logo" -> "zukunftsfond",
+        "sponsoring_slogan" -> "Winamp - it really whips the llama's ass",
+        "sponsoring_hidden" -> "true"
+      )))
+      rawToApiConverter.mergeSitebuildings(channel, master) mustBe RawChannelSiteBuilding(fields = Some(Map(
+        "sponsoring_slogan" -> "Winamp - it really whips the llama's ass",
+        "sponsoring_hidden" -> "true",
+        "header_logo" -> "zukunftsfond"
+      )))
+    }
+
+    "add elements from master if channel itself has no elements" in {
+      val channel = RawChannelSiteBuilding(elements = None)
+      val master = RawChannelSiteBuilding(elements = Some(Seq(RawElement())), fields = Some(Map.empty))
+      rawToApiConverter.mergeSitebuildings(channel, master) mustBe master
+    }
+
+    "add sub navi from master if channel itself has no sub navi" in {
+      val channel = RawChannelSiteBuilding(elements = None)
+      val master = RawChannelSiteBuilding(sub_navigation = Some(Seq(RawSectionReference())), fields = Some(Map.empty))
+      rawToApiConverter.mergeSitebuildings(channel, master) mustBe master
+    }
+
   }
 
-  "Site Building" must {
-
-    "not be inherited from parent that has a site building configuration but is not master" in new SiteBuildingTreeScope {
-      converter.calculateSiteBuilding(node2000) must not be converter.calculateSiteBuilding(node200)
-    }
-
-    "be inherited from next master even if the master has an empty site building configuration" in new SiteBuildingTreeScope {
-      converter.calculateSiteBuilding(node2000) mustBe empty
-    }
-
-    "be empty for a first level section that has no site building configuration (no inheritance from root)" in new SiteBuildingTreeScope {
-      converter.calculateSiteBuilding(node30) mustBe empty
-    }
-  }
 }